requirements.txt
Django==1.8.4
pep8==1.6.2
psycopg2==2.6.1
strictdict3==0.4.2
